feat: Implement HomePage component with data fetching and caching for leaderboard and stats
