+++ .gitignore
@@ -22,6 +22,7 @@ dist-ssr
 # Ignore backup artifacts from image optimizer
 public/*.bak-*
 *.bak-*
+public/.image-opt-manifest.json
 
 # Editor directories and files
 .vscode/*
